+++ lib/generators/clearance/install/templates/db/migrate/add_clearance_to_users.rb
@@ -9,6 +9,16 @@ class AddClearanceToUsers < ActiveRecord::Migration
 <% config[:new_indexes].values.each do |index| -%>
     <%= index %>
 <% end -%>
+
+    users = select_all('SELECT id FROM users WHERE remember_token IS NULL')
+
+    users.each do |user|
+      update <<-SQL
+        UPDATE users
+        SET remember_token = '#{Clearance::Token.new}'
+        WHERE id = '#{user['id']}'
+      SQL
+    end
   end
 
   def self.down